Change history and other UI improvements
Decoupled state from the the Board class, with State being a model
and Board remaining a view-controller.
Changes to the State are stored as instances derived from Delta and
can always be undo.
Added a undo/redo function
Added buttons to change the symbol under cursor, enabled only if the move
is directly permissible (doesn't guarantee it will lead to a solvable
state though).
Changed section colour generation to use HSL, which makes it possible to
generate complementary colours for the selector.